Rewrite CheckboxField and CheckboxGroupField
